requirements.txt
Pytorch == 1.4.0
scikit-learn == 0.23.2
numpy == 1.16.2
hnswlib == 0.4.0
networkx == 2.5
nni == 2.8